chore(ssz): Cleanup (#1612)
