+++ mod/primitives/pkg/merkle/hasher.go
@@ -21,22 +21,43 @@
 package merkle
 
 import (
-	"encoding/binary"
+	"runtime"
+	"unsafe"
 
+	"github.com/berachain/beacon-kit/mod/primitives/pkg/bytes"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/math"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/merkle/zero"
 	"github.com/prysmaticlabs/gohashtree"
+	"golang.org/x/sync/errgroup"
 )
+
+const (
+	// MinParallelizationSize is the minimum size of the input list that
+	// should be hashed using the default method. If the input list is smaller
+	// than this size, the overhead of parallelizing the hashing process is.
+	//
+	// TODO: This value is arbitrary and should be benchmarked to find the
+	// optimal value.
+	MinParallelizationSize = 5000
+	// two is a constant to make the linter happy.
+	two = 2
+)
+
+type MerkleHashFn[RootT ~[32]byte] func([]RootT, []RootT) error
 
 // Hasher can be re-used for constructing Merkle tree roots.
 type Hasher[RootT ~[32]byte] struct {
-	buffer Buffer[RootT]
+	// buffer is a reusable buffer for hashing.
+	buffer bytes.Buffer[RootT]
+	// hasher is the hashing function to use.
+	hasher MerkleHashFn[RootT]
 }
 
 // NewHasher creates a new merkle Hasher.
-func NewHasher[RootT ~[32]byte](buffer Buffer[RootT]) *Hasher[RootT] {
+func NewHasher[RootT ~[32]byte](buffer bytes.Buffer[RootT], hashFn MerkleHashFn[RootT]) *Hasher[RootT] {
 	return &Hasher[RootT]{
 		buffer: buffer,
+		hasher: hashFn,
 	}
 }
 
@@ -73,7 +94,7 @@ func (m *Hasher[RootT]) NewRootWithDepth(
 		}
 
 		newLayerSize := (layerLen + 1) / two
-		if err = BuildParentTreeRoots(buf[:newLayerSize], leaves); err != nil {
+		if err = m.hasher(buf[:newLayerSize], leaves); err != nil {
 			return zero.Hashes[depth], err
 		}
 		leaves, buf = buf[:newLayerSize], leaves
@@ -84,16 +105,85 @@ func (m *Hasher[RootT]) NewRootWithDepth(
 	return leaves[0], nil
 }
 
-// MixinLength takes a root element and mixes in the length of the elements
-// that were hashed to produce it.
+// BuildParentTreeRoots calls BuildParentTreeRootsWithNRoutines with the
+// number of routines set to runtime.GOMAXPROCS(0)-1.
+func BuildParentTreeRoots[RootT ~[32]byte](
+	outputList, inputList []RootT,
+) error {
+	err := BuildParentTreeRootsWithNRoutines(
+		//#nosec:G103 // on purpose.
+		*(*[][32]byte)(unsafe.Pointer(&outputList)),
+		//#nosec:G103 // on purpose.
+		*(*[][32]byte)(unsafe.Pointer(&inputList)),
+		runtime.GOMAXPROCS(0)-1,
+	)
+
+	// Convert out back to []RootT using unsafe pointer cas
+	return err
+}
+
+// BuildParentTreeRootsWithNRoutines optimizes hashing of a list of roots
+// using CPU-specific vector instructions and parallel processing. This
+// method adapts to the host machine's hardware for potential performance
+// gains over sequential hashing.
 //
-// TODO: move to ssz package.
-func MixinLength[RootT ~[32]byte](element RootT, length uint64) RootT {
-	chunks := make([][32]byte, two)
-	chunks[0] = element
-	binary.LittleEndian.PutUint64(chunks[1][:], length)
-	if err := gohashtree.Hash(chunks, chunks); err != nil {
-		return [32]byte{}
+// TODO: We do not use generics here due to the gohashtree library not
+// supporting generics.
+func BuildParentTreeRootsWithNRoutines(
+	outputList, inputList [][32]byte, n int,
+) error {
+	// Validate input list length.
+	inputLength := len(inputList)
+	if inputLength%2 != 0 {
+		return ErrOddLengthTreeRoots
 	}
-	return chunks[0]
+
+	// Build output variables
+	outputLength := inputLength / two
+
+	// If the input list is small, hash it using the default method since
+	// the overhead of parallelizing the hashing process is not worth it.
+	if inputLength < MinParallelizationSize {
+		//#nosec:G103 // used of unsafe calls should be audited.
+		return gohashtree.Hash(outputList, inputList)
+	}
+
+	// Otherwise parallelize the hashing process for large inputs.
+	// Take the max(n, 1) to prevent division by 0.
+	groupSize := inputLength / (two * max(n, 1))
+	twiceGroupSize := two * groupSize
+	eg := new(errgroup.Group)
+
+	// if n is 0 the parallelization is disabled and the whole inputList is
+	// hashed in the main goroutine at the end of this function.
+	for j := 0; j <= n; j++ {
+		eg.Go(func() error {
+			// inputList:  [-------------------2*groupSize-------------------]
+			//              ^                  ^                    ^        ^
+			//              |                  |                    |        |
+			// j*2*groupSize   (j+1)*2*groupSize    (j+2)*2*groupSize   End
+			//
+			// outputList: [---------groupSize---------]
+			//              ^                         ^
+			//              |                         |
+			//             j*groupSize         (j+1)*groupSize
+			//
+			// Each goroutine processes a segment of inputList that is twice as
+			// large as the segment it fills in outputList. This is because the
+			// hash
+			// operation reduces the
+			// size of the input by half.
+			// Define the segment of the inputList each goroutine will process.
+			segmentStart := j * twiceGroupSize
+			segmentEnd := min((j+1)*twiceGroupSize, inputLength)
+
+			return gohashtree.Hash(
+				outputList[j*groupSize:min((j+1)*groupSize, outputLength)],
+				inputList[segmentStart:segmentEnd],
+			)
+		})
+	}
+
+	// Wait for all goroutines to complete.
+	return eg.Wait()
 }
